more on add ingredient

+++ src/main/java/com/rihanna/neo4j/eg3/controller/model/RecipeModelAttribute.java
@@ -9,14 +9,14 @@ public class RecipeModelAttribute {
 	private List<String> includedIngredients;
 	private List<String> excludedIngredients;
 	private List<String> tags;
-	private Integer minCalory;
-	private Integer maxCalory;
-	private Integer minCarb;
-	private Integer maxCarb;
-	private Integer minFat;
-	private Integer maxFat;
-	private Integer minProtein;
-	private Integer maxProtein;
+	private Float minCalory;
+	private Float maxCalory;
+	private Float minCarb;
+	private Float maxCarb;
+	private Float minFat;
+	private Float maxFat;
+	private Float minProtein;
+	private Float maxProtein;
 	
 	public String getRecipeName() {
 		return recipeName;
@@ -48,52 +48,52 @@ public class RecipeModelAttribute {
 	public void setTags(List<String> tags) {
 		this.tags = tags;
 	}
-	public Integer getMinCalory() {
+	public Float getMinCalory() {
 		return minCalory == null ? 0 : minCalory;
 	}
-	public void setMinCalory(Integer minCalory) {
+	public void setMinCalory(Float minCalory) {
 		this.minCalory = minCalory;
 	}
-	public Integer getMaxCalory() {
+	public Float getMaxCalory() {
 		return maxCalory == null ? 0 : maxCalory;
 	}
-	public void setMaxCalory(Integer maxCalory) {
+	public void setMaxCalory(Float maxCalory) {
 		this.maxCalory = maxCalory;
 	}
-	public Integer getMinCarb() {
+	public Float getMinCarb() {
 		return minCarb == null ? 0 : minCarb;
 	}
-	public void setMinCarb(Integer minCarb) {
+	public void setMinCarb(Float minCarb) {
 		this.minCarb = minCarb;
 	}
-	public Integer getMaxCarb() {
+	public Float getMaxCarb() {
 		return maxCarb == null ? 0 : maxCarb;
 	}
-	public void setMaxCarb(Integer maxCarb) {
+	public void setMaxCarb(Float maxCarb) {
 		this.maxCarb = maxCarb;
 	}
-	public Integer getMinFat() {
+	public Float getMinFat() {
 		return minFat == null ? 0 : minFat;
 	}
-	public void setMinFat(Integer minFat) {
+	public void setMinFat(Float minFat) {
 		this.minFat = minFat;
 	}
-	public Integer getMaxFat() {
+	public Float getMaxFat() {
 		return maxFat == null ? 0 : maxFat;
 	}
-	public void setMaxFat(Integer maxFat) {
+	public void setMaxFat(Float maxFat) {
 		this.maxFat = maxFat;
 	}
-	public Integer getMinProtein() {
+	public Float getMinProtein() {
 		return minProtein == null ? 0 : minProtein;
 	}
-	public void setMinProtein(Integer minProtein) {
+	public void setMinProtein(Float minProtein) {
 		this.minProtein = minProtein;
 	}
-	public Integer getMaxProtein() {
+	public Float getMaxProtein() {
 		return maxProtein == null ? 0 : maxProtein;
 	}
-	public void setMaxProtein(Integer maxProtein) {
+	public void setMaxProtein(Float maxProtein) {
 		this.maxProtein = maxProtein;
 	}
