+++ src/main/java/com/rihanna/neo4j/eg3/dto/NutritionalValueDTO.java
@@ -2,15 +2,15 @@ package com.rihanna.neo4j.eg3.dto;
 
 public class NutritionalValueDTO {
 
-	private Integer calory;
-	private Integer carbs;
-	private Integer fat;
-	private Integer protein;
+	private Float calory;
+	private Float carbs;
+	private Float fat;
+	private Float protein;
 	
 	public NutritionalValueDTO() {	}
 
 	
-	public NutritionalValueDTO(Integer calory, Integer carbs, Integer fat, Integer protein) {
+	public NutritionalValueDTO(Float calory, Float carbs, Float fat, Float protein) {
 		super();
 		this.calory = calory;
 		this.carbs = carbs;
@@ -18,35 +18,35 @@ public class NutritionalValueDTO {
 		this.protein = protein;
 	}
 
-	public void addValues(Integer calory, Integer carbs, Integer fat, Integer protein) {
+	public void addValues(Float calory, Float carbs, Float fat, Float protein) {
 		this.calory += calory;
 		this.carbs += carbs;
 		this.fat += fat;
 		this.protein += protein;
 	}
 	
-	public Integer getCalory() {
+	public Float getCalory() {
 		return calory == null ? 0 : calory;
 	}
-	public void setCalory(Integer calory) {
+	public void setCalory(Float calory) {
 		this.calory = calory;
 	}
-	public Integer getCarbs() {
+	public Float getCarbs() {
 		return carbs == null ? 0 : carbs;
 	}
-	public void setCarbs(Integer carbs) {
+	public void setCarbs(Float carbs) {
 		this.carbs = carbs;
 	}
-	public Integer getFat() {
+	public Float getFat() {
 		return fat == null ? 0 : fat;
 	}
-	public void setFat(Integer fat) {
+	public void setFat(Float fat) {
 		this.fat = fat;
 	}
-	public Integer getProtein() {
+	public Float getProtein() {
 		return protein == null ? 0 : protein;
 	}
-	public void setProtein(Integer protein) {
+	public void setProtein(Float protein) {
 		this.protein = protein;
 	}
 }
